Applied HARMONY-3346 [drlvm][winx64][build] Warning was introduced into classloader.cpp
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@516421 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/vmcore/src/class_support/classloader.cpp
@@ -239,7 +239,7 @@ Class* ClassLoader::DefineClass(Global_Env* env, const char* class_name,
     }
     if (LookupClass(className) != NULL) {
         static const char* mess_templ = "Illegal attempt to redefine class : %s";
-        unsigned mess_size = strlen(mess_templ) + className->len + 1;
+        size_t mess_size = strlen(mess_templ) + className->len + 1;
         char* err_mess = (char*)STD_ALLOCA(mess_size);
         sprintf(err_mess, mess_templ, className->bytes);
         exn_raise_by_name("java/lang/LinkageError", err_mess);
